Highlight Paste Raw Text When User Clicks Textarea
When a paste is displayed, it is displayed in a formatted manner, followed by the text of the paste in raw ASCII. Often, the raw ASCII is used when copying and pasting text that you do not want non-ASCII text carrying over (e.g. UTF-8 symbols). To help with UX, the textarea was updated to highlight all text when a user clicks it to obtain DOM focus

Issues: #32